Graceful redis-less degradation. Comments and bugfixes.

+++ apps/dashboards/tests/test_cron.py
@@ -20,23 +20,36 @@ def _add_vote_in_past(rev, vote, days_back):
     v.save()
 
 
+def _make_backdated_revision(backdate):
+    r = revision(save=True)
+    r.created = r.created - timedelta(days=backdate)
+    r.save()
+    return r
+
+
 class TopUnhelpfulArticlesTests(TestCase):
     def test_no_old_articles(self):
+        """Make sure _get_old_articles() returns nothing with no votes."""
         result = _get_old_unhelpful()
         eq_(0, len(result))
 
     def test_no_current_articles(self):
+        """Make sure _get_current_articles() returns nothing with no votes."""
         result = _get_current_unhelpful({})
         eq_(0, len(result))
 
     def test_old_articles(self):
-        r = revision(save=True)
-        r.created = r.created - timedelta(days=90)
-        r.save()
+        """
+        Make sure _get_old_unhelpful() returns an unhelpful (no > yes) with
+        votes within time range.
+        """
+        r = _make_backdated_revision(90)
 
+        # Add 4 no votes 1.5 months ago
         for x in range(0, 4):
             _add_vote_in_past(r, 0, 45)
 
+        # Add 1 yes vote 1.5 months ago
         _add_vote_in_past(r, 1, 45)
 
         result = _get_old_unhelpful()
@@ -45,9 +58,11 @@ class TopUnhelpfulArticlesTests(TestCase):
         eq_(5, result[r.id]['total'])
 
     def test_old_articles_helpful(self):
-        r = revision(save=True)
-        r.created = r.created - timedelta(days=90)
-        r.save()
+        """
+        Make sure _get_old_unhelpful() doesn't return a helpful (no < yes)
+        with votes within time range.
+        """
+        r = _make_backdated_revision(90)
 
         for x in range(0, 4):
             _add_vote_in_past(r, 1, 45)
@@ -58,9 +73,11 @@ class TopUnhelpfulArticlesTests(TestCase):
         eq_(0, len(result))
 
     def test_current_articles(self):
-        r = revision(save=True)
-        r.created = r.created - timedelta(days=90)
-        r.save()
+        """
+        Make sure _get_current_unhelpful() returns an unhelpful (no > yes) with
+        votes within time range.
+        """
+        r = _make_backdated_revision(90)
 
         for x in range(0, 3):
             _add_vote_in_past(r, 0, 15)
@@ -78,9 +95,11 @@ class TopUnhelpfulArticlesTests(TestCase):
         eq_(5, result[r.id]['total'])
 
     def test_current_articles_helpful(self):
-        r = revision(save=True)
-        r.created = r.created - timedelta(days=90)
-        r.save()
+        """
+        Make sure _get_current_unhelpful() doesn't return a helpful (no < yes)
+        with votes within time range.
+        """
+        r = _make_backdated_revision(90)
 
         for x in range(0, 3):
             _add_vote_in_past(r, 1, 15)
@@ -94,9 +113,10 @@ class TopUnhelpfulArticlesTests(TestCase):
         eq_(0, len(result))
 
     def test_current_articles_not_in_old(self):
-        r = revision(save=True)
-        r.created = r.created - timedelta(days=90)
-        r.save()
+        """
+        Unhelpful articles in current but not in old shouldn't break stuff.
+        """
+        r = _make_backdated_revision(90)
 
         for x in range(0, 3):
             _add_vote_in_past(r, 0, 15)
@@ -118,6 +138,8 @@ class TopUnhelpfulArticlesCronTests(TestCase):
         super(TopUnhelpfulArticlesCronTests, self).setUp()
         try:
             self.redis = redis_client('helpfulvotes')
+            if self.redis is None:
+                raise SkipTest
             self.redis.flushdb()
             self.REDIS_KEY = settings.HELPFULVOTES_UNHELPFUL_KEY
         except (KeyError, AttributeError):
@@ -131,13 +153,13 @@ class TopUnhelpfulArticlesCronTests(TestCase):
         super(TopUnhelpfulArticlesCronTests, self).tearDown()
 
     def test_no_articles(self):
+        """Full cron with no articles returns no unhelpful articles."""
         cache_most_unhelpful_kb_articles()
         eq_(0, self.redis.llen(self.REDIS_KEY))
 
     def test_caching_unhelpful(self):
-        r = revision(save=True)
-        r.created = r.created - timedelta(days=90)
-        r.save()
+        """Cron should get the unhelpful articles."""
+        r = _make_backdated_revision(90)
 
         for x in range(0, 3):
             _add_vote_in_past(r, 0, 15)
@@ -149,12 +171,12 @@ class TopUnhelpfulArticlesCronTests(TestCase):
 
         eq_(1, self.redis.llen(self.REDIS_KEY))
         result = self.redis.lrange(self.REDIS_KEY, 0, 1)
-        eq_('%d:%.1f:%.1f:%.1f:%.1f' % (r.id, 5.0, 0.4, 0.0, 0.0), result[0])
+        eq_(u'%d::%.1f::%.1f::%.1f::%.1f::%s::%s' % (r.id, 5.0, 0.4, 0.0, 0.0,
+                                r.document.slug, r.document.title), unicode(result[0], "utf-8"))
 
     def test_caching_helpful(self):
-        r = revision(save=True)
-        r.created = r.created - timedelta(days=90)
-        r.save()
+        """Cron should ignore the helpful articles."""
+        r = _make_backdated_revision(90)
 
         for x in range(0, 3):
             _add_vote_in_past(r, 1, 15)
@@ -167,9 +189,8 @@ class TopUnhelpfulArticlesCronTests(TestCase):
         eq_(0, self.redis.llen(self.REDIS_KEY))
 
     def test_caching_changed_helpfulness(self):
-        r = revision(save=True)
-        r.created = r.created - timedelta(days=90)
-        r.save()
+        """Changed helpfulness should be calculated correctly."""
+        r = _make_backdated_revision(90)
 
         for x in range(0, 4):
             _add_vote_in_past(r, 0, 45)
@@ -187,14 +208,13 @@ class TopUnhelpfulArticlesCronTests(TestCase):
 
         eq_(1, self.redis.llen(self.REDIS_KEY))
         result = self.redis.lrange(self.REDIS_KEY, 0, 1)
-        eq_('%d:%.1f:%.1f:%.1f:%.1f' % (r.id, 5.0, 0.4, 0.2, 0.0), result[0])
+        eq_(u'%d::%.1f::%.1f::%.1f::%.1f::%s::%s' % (r.id, 5.0, 0.4, 0.2, 0.0,
+                                 r.document.slug, r.document.title), unicode(result[0], "utf-8"))
 
     def test_caching_sorting(self):
         """Tests if Bayesian Average sorting works correctly."""
         # This should be at the bottom.
-        r = revision(save=True)
-        r.created = r.created - timedelta(days=90)
-        r.save()
+        r = _make_backdated_revision(90)
 
         for x in range(0, 26):
             _add_vote_in_past(r, 1, 15)
@@ -203,9 +223,7 @@ class TopUnhelpfulArticlesCronTests(TestCase):
             _add_vote_in_past(r, 0, 15)
 
         # This should be at the top.
-        r2 = revision(save=True)
-        r2.created = r2.created - timedelta(days=90)
-        r2.save()
+        r2 = _make_backdated_revision(90)
 
         for x in range(0, 61):
             _add_vote_in_past(r2, 1, 15)
@@ -214,9 +232,7 @@ class TopUnhelpfulArticlesCronTests(TestCase):
             _add_vote_in_past(r2, 0, 15)
 
         # This should be in the middle.
-        r3 = revision(save=True)
-        r3.created = r3.created - timedelta(days=90)
-        r3.save()
+        r3 = _make_backdated_revision(90)
 
         for x in range(0, 31):
             _add_vote_in_past(r3, 1, 15)
@@ -228,6 +244,6 @@ class TopUnhelpfulArticlesCronTests(TestCase):
 
         eq_(3, self.redis.llen(self.REDIS_KEY))
         result = self.redis.lrange(self.REDIS_KEY, 0, 3)
-        assert '%d:%.1f:' % (r2.id, 242.0) in result[0]
-        assert '%d:%.1f:' % (r3.id, 122.0) in result[1]
-        assert '%d:%.1f:' % (r.id, 102.0) in result[2]
+        assert '%d::%.1f:' % (r2.id, 242.0) in result[0]
+        assert '%d::%.1f:' % (r3.id, 122.0) in result[1]
+        assert '%d::%.1f:' % (r.id, 102.0) in result[2]
